+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 -----------------
 
 * Added documentation for the default schemas.
+* Change the implementation to check what settings form was saved.
+* Added new util method 'set_setting' where value transformations are handled.
 
 
 0.1b4(2013-05-02)